Moved tests per Perth's comment
